Add @avalance123's magento-twilio extension as a planned feature.
Rather than completely re-invent the wheel, let's rather add the extension from Avalance123 to this extension.  His extension has not been updated in a while and will therefore need some work to make sure that it works and is still compatible with the various editions since its release 3 years ago.